refactor(pinocchio): rename torque_fitting.py to poly_torque_util.py
Pure file rename. The module is a 1D polyfit utility, not a swing-fit
driver, so the new name reflects intent.

- git mv module and both test files (history preserved)
- update imports in tests to use the new module path

No behavior changes.

Closes #4135
